+++ packages/powersync-op-sqlite/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @powersync/op-sqlite
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [77a9ed2]
+  - @powersync/common@1.22.0
+
+## 0.1.2
+
+### Patch Changes
+
+- 7c9c41d: Update op-sqlite to v10.1.0 for compatibility with React Native >0.76
+
 ## 0.1.1
 
 ### Patch Changes