+++ tools/diagnostics-app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # diagnostics-app
 
+## 0.7.4
+
+### Patch Changes
+
+- @powersync/web@1.12.1
+
+## 0.7.3
+
+### Patch Changes
+
+- Updated dependencies [7e23d65]
+- Updated dependencies [36af0c8]
+  - @powersync/web@1.12.0
+
 ## 0.7.2
 
 ### Patch Changes